(BETA) Add global symbol and currency conversions
This still needs tests written but is at a point to be peer reviewed. The goal is to get a way to get a standard symbol (market) or currency that can be a common global string and then converted back for each individual exchange. This allows databases, etc. to be built on ExchangeSharp a lot easier.

Please send any feedback.