Switch web socket to raw bytes
Some exchanges gzip compress the bytes, and this needs to be handled in the exchange level API as .NET does not provide an automatic compression handler for web sockets. This could be added automatically later as an option, but for now it is up to each exchange api to handle it.