fix GetFullOrderBookWebSocketAsync data inconsistency
Proposal for removing maxCount parameter from ParseOrderBookFromJTokenArrays and ParseOrderBookFromJTokenDictionaries.
Currently we are trimming the incoming order book data before processing it, which seems wrong and causes order book inconsistency as mentioned in #713 .
Instead I propose moving maxCount trimming logic to GetFullOrderBookWebSocketAsync, because that's the method responsible for order book data store as well.